% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pkgs_info.R
\name{get_pkgs_info}
\alias{get_pkgs_info}
\title{Get information about packages used in a project}
\usage{
get_pkgs_info(
  all.pkgs = TRUE,
  out.dir = getwd(),
  bib.file = "grateful-refs.bib",
  include.RStudio = FALSE,
  ...
)
}
\arguments{
\item{all.pkgs}{Logical. Include all packages used in scripts within
the project/folder (the default), or only packages used in the current session?
If TRUE, uses \code{\link[renv]{dependencies}},
otherwise uses \code{\link[utils]{sessionInfo}}.}

\item{out.dir}{Directory to save the BibTeX file with the references.
Defaults to working directory.}

\item{bib.file}{Name of the file to save the BibTeX references.}

\item{include.RStudio}{Logical. If TRUE, adds a citation for the current
version of RStudio.}

\item{...}{
  Arguments passed on to \code{\link[=scan_packages]{scan_packages}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
A data.frame with package info,
and a file containing package references in BibTeX format.
}
\description{
This function scans the project for R packages used and returns a data frame
with package names, version, and citation keys.
}
\examples{
\dontrun{
get_pkgs_info()
}
}
