% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grateful-package.R
\docType{package}
\name{grateful-package}
\alias{grateful}
\alias{grateful-package}
\title{grateful: Facilitate Citation of R Packages}
\description{
Facilitates the citation of R packages used in analysis projects. Scans project for packages used, gets their citations, and produces a document with citations in the preferred bibliography format, ready to be pasted into reports or manuscripts. Alternatively, 'grateful' can be used directly within an 'R Markdown' or 'Quarto' document.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pakillo.github.io/grateful/}
  \item Report bugs at \url{https://github.com/Pakillo/grateful/issues}
}

}
\author{
\strong{Maintainer}: Francisco Rodriguez-Sanchez \email{f.rodriguez.sanc@gmail.com} (\href{https://orcid.org/0000-0002-7981-1599}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Connor P. Jackson \email{connor@cpjackson.net} (\href{https://orcid.org/0000-0002-4220-5210}{ORCID})
}

Other contributors:
\itemize{
  \item Shaurita D. Hutchins [contributor]
  \item James M. Clawson [contributor]
}

}
\keyword{internal}
