% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grateful-package.r
\docType{package}
\name{grateful}
\alias{grateful}
\title{grateful.}
\description{
The goal of `grateful` is to make it very easy to cite the R packages used in any report or publication. By calling a single function, it will scan the project for R packages used, and generate a document with citations in the desired output format (Word, PDF, HTML, Markdown).
}
\details{
Importantly, these references can be formatted for a specific journal so that we can just paste them directly into the bibliography list of our manuscript or report.

If you prefer to get just a BibTeX file with all package citations, it can also be done (see examples in https://github.com/Pakillo/grateful).
}
