% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cite_packages.R
\name{cite_packages}
\alias{cite_packages}
\title{Cite R packages used in a project}
\usage{
cite_packages(
  all.pkg = TRUE,
  style = NULL,
  out.format = "html",
  out.dir = getwd(),
  ...
)
}
\arguments{
\item{all.pkg}{Logical. Include all packages used in scripts within the
project/folder (the default), or only packages used in the current session?
If TRUE, uses \code{\link[renv]{dependencies}},
otherwise uses \code{\link[utils]{sessionInfo}}.}

\item{style}{Optional. Citation style to format references. See \url{https://www.zotero.org/styles}.}

\item{out.format}{Output format, either "docx", "pdf", "html", or "Rmd" (Rmarkdown).}

\item{out.dir}{Directory to save the output document and a BibTeX file with the references.
Defaults to working directory.}

\item{...}{Other parameters passed to \code{\link[renv]{dependencies}}.}
}
\value{
A document with rendered citations in the desired format,
plus a file ("pkg-refs.bib") containing references in BibTeX format.
}
\description{
This is a wrapper function.
}
\examples{
\dontrun{
library(grateful)
cite_packages()
cite_packages(style = "ecology", out.format = "docx")
}
}
