% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cite_packages.R
\name{cite_packages}
\alias{cite_packages}
\title{Cite R packages used in a project}
\usage{
cite_packages(generate.document = TRUE, all.pkg = TRUE,
  include.rmd = TRUE, style = NULL, out.format = "html",
  out.dir = getwd(), include.RStudio = FALSE, ...)
}
\arguments{
\item{generate.document}{Logical. If TRUE (default), generate a .Rmd file with the citations. Otherwise, simply build the .bib file and return the list of package citation keys.}

\item{all.pkg}{Logical. Include all packages used in scripts within the project/folder (the default), or only packages used in the current session? If TRUE, uses \code{\link[checkpoint]{scanForPackages}}, otherwise uses \code{\link[utils]{sessionInfo}}.}

\item{include.rmd}{Logical. Include packages used in Rmarkdown documents? (default is TRUE, requires \code{knitr} package).}

\item{style}{Optional. Citation style to format references. See \url{http://citationstyles.org/styles/}.}

\item{out.format}{Output format, either "docx" (Word), "pdf", "html", or "md" (markdown).}

\item{out.dir}{Directory to save the output document and a BibTeX file with the references. Defaults to working directory.}

\item{include.RStudio}{Logical. If TRUE, adds a citation for the current version of RStudio, if run within RStudio interactively.}

\item{...}{Other parameters passed to \code{\link[checkpoint]{scanForPackages}}.}
}
\value{
A document with rendered citations in the desired format, plus a file ("pkg-refs.bib") containing references in BibTeX format.
}
\description{
This is a wrapper function.
}
\examples{
\dontrun{
library(grateful)
cite_packages()
cite_packages(style = "ecology", out.format = "docx")
}
}
