% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rmd.R
\name{create_rmd}
\alias{create_rmd}
\title{Create template rmarkdown file with package citations}
\usage{
create_rmd(bib.list, bibfile = "pkg-refs.bib", csl = NULL,
  filename = "refs.Rmd")
}
\arguments{
\item{bib.list}{List of package citations in BibTeX format, as produced by \code{get_citations}.}

\item{bibfile}{Name of the file containing references in BibTeX format, as produced by \code{get_citations}.}

\item{csl}{Optional. Citation style to format references. See \url{http://citationstyles.org/styles/}.}

\item{filename}{Name of the rmarkdown file}
}
\value{
An rmarkdown file
}
\description{
Create template rmarkdown file with package citations
}
\examples{
library(grateful)
pkgs <- scan_packages()
cites <- get_citations(pkgs)
rmd <- create_rmd(cites)
}
