% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_citations.R
\name{get_citations}
\alias{get_citations}
\title{Get citations for packages}
\usage{
get_citations(
  pkgs,
  out.dir = getwd(),
  bib.file = "grateful-refs.bib",
  include.RStudio = FALSE
)
}
\arguments{
\item{pkgs}{Character vector of package names, e.g. as produced by
\code{scan_packages}.}

\item{out.dir}{Directory to save the BibTeX file with the references.
Defaults to working directory.}

\item{bib.file}{Name of the file to save the BibTeX references.}

\item{include.RStudio}{Logical. If TRUE, adds a citation for the current
version of RStudio.}
}
\value{
A file on the specified \code{out.dir} containing the package references in BibTeX
  format. If assigned a name, \code{get_citations} will also return a list with citation keys for
  each citation (without @).
}
\description{
Get citations for packages
}
\examples{
\dontrun{
citekeys <- get_citations(c("lme4", "mgcv"))

pkgs <- scan_packages()
citekeys <- get_citations(pkgs$pkg)
}
}
