% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_citations.R
\name{get_citations}
\alias{get_citations}
\title{Get citations for packages}
\usage{
get_citations(pkgs, filename = NULL)
}
\arguments{
\item{pkgs}{Character vector of package names, e.g. as produced by \code{scan_packages}.}

\item{filename}{Optional. Name of BibTeX file containing packages references.}
}
\value{
Nothing by default. If assigned a name, a list with package citations in BibTeX format. Optionally, a file with references in BibTeX format.
}
\description{
Get citations for packages
}
\examples{
library(grateful)
pkgs <- scan_packages()
cites <- get_citations(pkgs)
}
