% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_csl.R
\name{get_csl}
\alias{get_csl}
\title{Get a journal citation style from internet repository}
\usage{
get_csl(name)
}
\arguments{
\item{name}{Name of the journal, exactly as in \url{https://github.com/citation-style-language/styles}.}
}
\value{
The CSL file is saved in the working directory
}
\description{
Get a journal citation style from internet repository
}
\examples{
library(grateful)
get_csl("ecosistemas")
}
