% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pkgs_info.R
\name{get_pkgs_info}
\alias{get_pkgs_info}
\title{Get information about packages used in a project}
\usage{
get_pkgs_info(
  pkgs = "All",
  out.dir = NULL,
  omit = NULL,
  cite.tidyverse = TRUE,
  cite.grateful = FALSE,
  dependencies = FALSE,
  bib.file = "grateful-refs",
  include.RStudio = FALSE,
  ...
)
}
\arguments{
\item{pkgs}{Character. Either "All" to include all packages used in scripts
within the project/folder (the default), or "Session" to include only packages
used in the current session.
Alternatively, \code{pkgs} can also be a character vector of package names to
get citations for (see examples).}

\item{out.dir}{Directory to save the BibTeX file with references.
It is recommended to set \code{out.dir = getwd()}.}

\item{omit}{Character vector of package names to be omitted from the citation
report.}

\item{cite.tidyverse}{Logical. If \code{TRUE}, all tidyverse packages
(dplyr, ggplot2, etc) will be collapsed into a single citation
of the 'tidyverse', as recommended by the tidyverse team.}

\item{cite.grateful}{Logical. Include a citation to \code{grateful} package?
Default is FALSE.}

\item{dependencies}{Logical. Include the dependencies of your used packages?
If \code{TRUE}, will include all the packages that your used packages depend on.}

\item{bib.file}{Desired name for the BibTeX file containing packages references
("grateful-refs" by default).}

\item{include.RStudio}{Logical. If \code{TRUE}, adds a citation for the
current version of RStudio.}

\item{...}{Other parameters passed to \code{\link[renv:dependencies]{renv::dependencies()}}.}
}
\value{
A data.frame with package info,
and a file containing package references in BibTeX format.
}
\description{
This function scans the project for R packages used, saves a BibTeX file with
package references, and returns a data frame with package names, version,
and citation keys.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_pkgs_info(out.dir = tempdir())
get_pkgs_info(pkgs = c("renv", "remotes"), out.dir = tempdir())
\dontshow{\}) # examplesIf}
}
