% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pkgs_info.R
\name{get_pkgs_info}
\alias{get_pkgs_info}
\title{Get information about packages used in a project}
\usage{
get_pkgs_info(
  pkgs = "All",
  cite.tidyverse = FALSE,
  dependencies = FALSE,
  out.dir = getwd(),
  bib.file = "grateful-refs.bib",
  include.RStudio = FALSE,
  ...
)
}
\arguments{
\item{pkgs}{Character. Either "All" to include all packages used in scripts within
the project/folder (the default), or "Session" to include only packages
used in the current session.
\code{pkgs} can also be a character vector of package names to get citations for
(see examples).}

\item{cite.tidyverse}{Logical. If \code{TRUE}, all tidyverse packages (dplyr, ggplot2, etc)
will be collapsed into a single citation of the 'tidyverse'.}

\item{dependencies}{Logical. Include the dependencies of your used packages?
If \code{TRUE}, will include all the packages that your used packages depend on.}

\item{out.dir}{Directory to save the BibTeX file with the references.
Defaults to working directory.}

\item{bib.file}{Name of the file to save the BibTeX references.}

\item{include.RStudio}{Logical. If TRUE, adds a citation for the current
version of RStudio.}

\item{...}{
  Arguments passed on to \code{\link[=scan_packages]{scan_packages}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
A data.frame with package info,
and a file containing package references in BibTeX format.
}
\description{
This function scans the project for R packages used and returns a data frame
with package names, version, and citation keys.
}
\examples{
\dontrun{
get_pkgs_info()
get_pkgs_info(pkgs = c("lme4", "mgcv"))
}
}
