% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grateful-package.r
\docType{package}
\name{grateful}
\alias{grateful}
\title{Grateful: Facilitate citation of R Packages}
\description{
Find R packages used in a project, create a BibTeX file of citations, and
cite them in either a new or existing document. See the primary function
\code{\link[=cite_packages]{cite_packages()}} for full details and usage instructions.
}
