% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_packages.R
\name{scan_packages}
\alias{scan_packages}
\title{Scan a project or folder for packages used}
\usage{
scan_packages(all.pkgs = TRUE, ...)
}
\arguments{

\item{all.pkgs}{Logical. Include all packages used in scripts within
the project/folder (the default), or only packages used in the current session?
If TRUE, uses \code{\link[renv]{dependencies}},
otherwise uses \code{\link[utils]{sessionInfo}}.}

\item{...}{Other parameters passed to \code{\link[renv]{dependencies}}.}

}
\value{
a character vector of package names
}
\description{
Scan a project or folder for packages used
}
\examples{
library(grateful)
scan_packages()
}
