% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_packages.R
\name{scan_packages}
\alias{scan_packages}
\title{Scan a project or folder for packages used}
\usage{
scan_packages(all.pkgs = TRUE, include.Rmd = FALSE, ...)
}
\arguments{
\item{all.pkgs}{Logical. Include all packages used in scripts within the
project/folder (the default), or only packages used in the current session?
If TRUE, uses \code{\link[checkpoint]{scanForPackages}}, otherwise uses
\code{\link[utils]{sessionInfo}}.}

\item{include.Rmd}{Logical. Include packages used in Rmarkdown documents?
(default is TRUE, requires \code{knitr} package).}

\item{...}{Other parameters passed to
\code{\link[checkpoint]{scanForPackages}}.}
}
\value{
a character vector of package names
}
\description{
Scan a project or folder for packages used
}
\examples{
library(grateful)
scan_packages()
}
