% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_packages.R
\name{scan_packages}
\alias{scan_packages}
\title{Scan a project or folder for packages used}
\usage{
scan_packages(pkgs = "All", ...)
}
\arguments{
\item{pkgs}{Character. Either "All" to include all packages used in scripts within
the project/folder (the default), or "Session" to include only packages
used in the current session.
\code{pkgs} can also be a character vector of package names to get citations for
(see examples).}

\item{...}{Other parameters passed to \code{\link[renv]{dependencies}}.}
}
\value{
a data.frame with package names and versions
}
\description{
Scan a project or folder for packages used
}
\examples{
scan_packages()
scan_packages(pkgs = "Session")
scan_packages(pkgs = c("lme4", "vegan", "mgcv"))
}
